git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/branches/sgopt@5336 4eea3252-f0fb-4393-894d-40516dce545b

+++ optimization/tests/testGridGeneration.cpp
@@ -75,7 +75,8 @@ BOOST_AUTO_TEST_CASE(TestIterativeGridGenerators) {
       return n * n;
     };
     gridGen.setAdaptivity(adaptivityFunction);
-    BOOST_CHECK_EQUAL(gridGen.getAdaptivity()(42), 42 * 42);
+    BOOST_CHECK_EQUAL(gridGen.getAdaptivity()(42),
+                      static_cast<size_t>(42 * 42));
   }
 
   for (auto& grid : grids) {
